--- demo/src/main/java/com/example/demo/product/ProductController.java
@@ -1,45 +1,0 @@
-package com.example.demo.product;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-import java.util.List;
-
-@RestController
-@RequestMapping(path = "/api/v1/product")
-public class ProductController {
-
-
-    private final ProductService productService;
-
-    @Autowired
-    public ProductController(ProductService productService) {
-        this.productService = productService;
-    }
-
-    @GetMapping(path = "{productId}")
-    public Product getProduct(@PathVariable("productId") Long productId) {
-        return productService.getProduct(productId);
-    }
-
-    @GetMapping
-    public List<ProductDTO> getAllProducts() {
-        return productService.getAllProducts();
-    }
-
-   @PostMapping
-    public void registerNewProduct(@RequestBody Product product) {
-        productService.addNewProducts(product);
-    }
-
-    @DeleteMapping(path = "{productId}")
-    public void deleteProduct(@PathVariable("productId") Long productId) {
-        productService.deleteProduct(productId);
-    }
-
-    @PutMapping(path = "{productId}")
-    public void updateProduct(@PathVariable("productId") Long productId
-            ,@RequestParam(required = false) String name
-            ,@RequestParam(required = false) String description)
-    {
-        productService.updateProduct(productId, name, description);
-    }
-}
