--- demo/src/main/java/com/example/demo/product/ProductDTO.java
@@ -1,5 +1,0 @@
-package com.example.demo.product;
-
-public record ProductDTO(Long id, String name, double price) {
-
-}
