--- demo/src/main/java/com/example/demo/product/ProductDTOMapper.java
@@ -1,18 +1,0 @@
-package com.example.demo.product;
-
-import org.springframework.stereotype.Service;
-
-import java.util.function.Function;
-
-@Service
-public class ProductDTOMapper implements Function<Product, ProductDTO> {
-
-    @Override
-    public ProductDTO apply(Product product) {
-        return new ProductDTO(
-                product.getId(),
-                product.getName(),
-                product.getPrice()
-        );
-    }
-}
