+++ demo/src/main/java/com/example/demo/product/modelMapper/ProductDTO.java
@@ -1,0 +1,37 @@
+package com.example.demo.product.modelMapper;
+
+public class ProductDTO {
+
+    private Long id;
+    private String name;
+    private double price;
+
+
+    //работи и без конструктора.
+    public ProductDTO() {
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public double getPrice() {
+        return price;
+    }
+
+    public void setPrice(double price) {
+        this.price = price;
+    }
+}
