+++ demo/src/main/java/com/example/demo/product/repostory/ProductRepository.java
@@ -1,0 +1,16 @@
+package com.example.demo.product.repostory;
+import com.example.demo.product.productConfiguration.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+import java.util.Optional;
+
+
+@Repository
+public interface ProductRepository
+        extends JpaRepository<Product, Long> {
+
+    @Query("SELECT name FROM Product products WHERE products.name = ?1")
+    Optional<Product> findByProductByName(String name);
+
+}
